requirement.txt
pymongo
python-dotenv
requests
fastapi
uvicorn
